build author blades based on urls in block content
