+++ scripts/scripts.js
@@ -16,9 +16,6 @@ const LCP_BLOCKS = []; // add your LCP blocks to the list
 
 const TEMPLATE_LIST = ['category', 'article'];
 
-const DEFAULT_CATEGORY_PATH = '/news';
-const DEFAULT_CATEGORY_NAME = 'News';
-
 /**
  * load fonts.css and set a session storage flag
  */
@@ -169,191 +166,6 @@ async function loadPage() {
   await loadLazy(document);
   await createContentAndAdsSections(document);
   loadDelayed();
-}
-
-/**
- * Returns the true origin of the current page in the browser.
- * If the page is running in a iframe with srcdoc, the ancestor origin is returned.
- * @returns {String} The true origin
- */
-export function getOrigin() {
-  const { location } = window;
-  return location.href === 'about:srcdoc' ? window.parent.location.origin : location.origin;
-}
-
-/**
- * Returns the true of the current page in the browser.mac
- * If the page is running in a iframe with srcdoc,
- * the ancestor origin + the path query param is returned.
- * @returns {String} The href of the current page or the href of the block running in the library
- */
-export function getHref() {
-  if (window.location.href !== 'about:srcdoc') return window.location.href;
-
-  const { location: parentLocation } = window.parent;
-  const urlParams = new URLSearchParams(parentLocation.search);
-  return `${parentLocation.origin}${urlParams.get('path')}`;
-}
-
-/**
- * Returns a picture element with webp and fallbacks
- * @param {string} src The image URL
- * @param {string} [alt] The image alternative text
- * @param {boolean} [eager] Set loading attribute to eager
- * @param {Array} [breakpoints] Breakpoints and corresponding params (eg. width)
- * @returns {Element} The picture element
- */
-export function createOptimizedPicture(src, alt = '', eager = false, breakpoints = [{ media: '(min-width: 600px)', width: '2000' }, { width: '750' }]) {
-  const url = new URL(src, getHref());
-  const picture = document.createElement('picture');
-  const { pathname } = url;
-  const ext = pathname.substring(pathname.lastIndexOf('.') + 1);
-
-  // webp
-  breakpoints.forEach((br) => {
-    const source = document.createElement('source');
-    if (br.media) source.setAttribute('media', br.media);
-    source.setAttribute('type', 'image/webp');
-    source.setAttribute('srcset', `${pathname}?width=${br.width}&format=webply&optimize=medium`);
-    picture.appendChild(source);
-  });
-
-  // fallback
-  breakpoints.forEach((br, i) => {
-    if (i < breakpoints.length - 1) {
-      const source = document.createElement('source');
-      if (br.media) source.setAttribute('media', br.media);
-      source.setAttribute('srcset', `${pathname}?width=${br.width}&format=${ext}&optimize=medium`);
-      picture.appendChild(source);
-    } else {
-      const img = document.createElement('img');
-      img.setAttribute('loading', eager ? 'eager' : 'lazy');
-      img.setAttribute('alt', alt);
-      picture.appendChild(img);
-      img.setAttribute('src', `${pathname}?width=${br.width}&format=${ext}&optimize=medium`);
-    }
-  });
-
-  return picture;
-}
-
-/**
- * @typedef QueryIndexRecord
- * @property {string} path Absolute path of the page on the site.
- * @property {string} title Title of the page from its metadata.
- * @property {string} description Summary description of the page from its metadata.
- * @property {string} image Absolute URL of the image from the page's metadata.
- * @property {string} category Name of the page's category.
- * @property {string} template Template assigned to the page associated with the index record.
- * @property {string} author Name of the author associated with the index record.
- * @property {string} authorimage Relative URL of the author's picture.
- * @property {string} authortitle Job title of the record's author.
- * @property {string} authordescription Bio information for the record's author.
- * @property {string} publisheddate Full, human-readable date when the record was published.
- * @property {string} keywords Comma-separated list of keywords associated with the index record.
- * @property {string} company-names Comma-separated list of companies to which the index record
- *  applies.
- * @property {string} company-webpages Comma-separated list of websites for the index's companies.
- * @property {string} lastModified Unix timestamp of the last time the index record was modified.
- */
-
-let cachedIndex;
-/**
- * Queries the site's index and only includes those records that match a given filter.
- * @param {function} filter Filter through which each record will be sent. The function will
- *  receive a single argument: the current record's raw data. The function should return true
- *  to include the record in the final result, or false to exclude it from the result.
- * @returns {Promise<Array<QueryIndexRecord>>} Resolves with an array of information for all
- *  matching records.
- */
-export async function queryIndex(filter) {
-  let index = cachedIndex;
-  if (!index) {
-    // will need to be updated to use ffetch for performance.
-    const res = await fetch('/query-index.json');
-    if (res && res.ok) {
-      try {
-        index = await res.json();
-        // storing the query index in memory as a speed performance optimization.
-        // may need to revisit this if the query index gets very large.
-        cachedIndex = index;
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.log('Unable to parse query index json', e);
-      }
-    }
-  }
-  const { data = [] } = index || {};
-  return data.filter(filter);
-}
-
-/**
- * Retrieves all articles whose path matches one of a given list of path values.
- * @param {Array<string>} paths List of all article paths to retrieve from the site's collection
- *  of articles.
- * @returns {Promise<Array<QueryIndexRecord>>} Resolves with an array of information for all
- *  matching articles.
- */
-export async function getArticlesByPath(paths) {
-  const pathLookup = {};
-  paths.forEach((path) => { pathLookup[path] = true; });
-  return queryIndex((record) => !!pathLookup[record.path]);
-}
-
-/**
- * Retrieves all authors whose full name matches one of a given list of values.
- * @param {Array<string>} names Names of authors to retrieve from the site's collection
- *  of authors.
- * @returns {Promise<Array<QueryIndexRecord>>} Resolves with an array of information for
- *  all matching authors.
- */
-export async function getAuthorsByName(names) {
-  const nameLookup = {};
-  names.forEach((name) => { nameLookup[name] = true; });
-  return queryIndex((record) => String(record.path).startsWith('/authors/') && !!nameLookup[record.author]);
-}
-
-/**
- * Retrieves all authors from the site.
- * @returns {Promise<Array<QueryIndexRecord>>} Resolves with an array of information for
- *  all authors on the site.
- */
-export async function getAllAuthors() {
-  return queryIndex((record) => String(record.path).startsWith('/authors/'));
-}
-
-/**
- * Retrieves the full, absolute path to a given path's category.
- * @param {string} path Path to a page on the site.
- * @returns {string} Path to the page's category.
- */
-export function getCategoryPath(path) {
-  if (!path || path === '/') {
-    return DEFAULT_CATEGORY_PATH;
-  }
-  let pathStr = String(path);
-  while (pathStr.length > 1 && pathStr.endsWith('/')) {
-    pathStr = pathStr.substring(0, pathStr.length - 1);
-  }
-  const lastSlash = pathStr.lastIndexOf('/');
-  if (lastSlash <= 0) {
-    // fall back to default category if unexpected path format
-    return DEFAULT_CATEGORY_PATH;
-  }
-  return pathStr.substring(0, lastSlash);
-}
-
-/**
- * Retrieves the category name of a record from the index.
- * @param {QueryIndexRecord} record Record from the site's query index.
- * @returns {string} Category name assigned to the record.
- */
-export function getCategoryName(record) {
-  const category = String(record.category).trim();
-  if (!category.length) {
-    return DEFAULT_CATEGORY_NAME;
-  }
-  return category;
 }
 
 loadPage();
